Extract reusable component MetricDetails
